proxy/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
litellm==1.56.5
python-dotenv==1.0.0
supabase==2.10.0
stripe==8.0.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
httpx==0.26.0
